feat(single): add higher-order lettable version of single
